requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
mock==3.0.5
npm==0.1.1
optional-django==0.1.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
